# core/src/main/resources/web/etc/swagger.yaml
swagger: "2.0"
info:
  description: "This is LEIA learns server. [Swagger contract](resources/etc/swagger.yaml)."
  version: 1.0.0
  title: LEIA learns
  termsOfService: http://swagger.io/terms/
  contact:
    url: http://leialearns.org/contact/
  license:
    name: LGPLv3
    url: https://www.gnu.org/licenses/lgpl-3.0.en.html
## host: localhost
basePath: /api/greeter
schemes:
- http
consumes: []
produces:
- application/json
paths:
  /api/state/step/action:
    post:
      tags:
      - StateTracker
      description: Record an action step
      operationId: recordActionStep
      parameters:
        - in: formData
          name: currentState
          description: The key of the model node that represents the current state
          type: string
        - in: formData
          name: vocabulary
          description: The key of the vocabulary
          type: string
        - in: formData
          name: symbol
          description: The name of the action symbol
          type: string
      responses:
        200:
          description: Action recorded
          schema:
            type: string
  /api/state/step/perception:
    post:
      tags:
        - StateTracker
      description: Record an perception step
      operationId: recordPerceptionStep
      parameters:
        - in: formData
          name: currentState
          description: The key of the model node that represents the current state
          type: string
        - in: formData
          name: vocabulary
          description: The key of the vocabulary
          type: string
        - in: formData
          name: symbol
          description: The name of the perception symbol
          type: string
      responses:
        200:
          description: Perception recorded
          schema:
            type: string
  /api/vocabulary/{key}:
    post:
      tags:
      - Vocabulary
      description: Create a Vocabulary for symbols
      operationId: createVocabulary
      parameters:
      - in: path
        name: key
        description: The key of the new Vocabulary
        required: true
        type: string
      responses:
        200:
          description: Successful operation
  /api/vocabulary:
    get:
      tags:
      - Vocabulary
      description: Get the keys of all known vocabularies
      operationId: getVocabularyKeys
      responses:
        200:
          description: Successful operation
          schema:
            $ref: "#/definitions/ArrayOfString"
      produces:
      - application/json
  /api/vocabulary/{key}/symbol/{symbol}:
    post:
      tags:
      - Vocabulary
      description: Get or create a Symbol in the given Vocabulary
      operationId: getOrCreateSymbol
      parameters:
      - in: path
        name: key
        description: The key of an existing Vocabulary
        required: true
        type: string
      - in: path
        name: symbol
        description: The name of the new Symbol
        required: true
        type: string
      responses:
        200:
          description: Successful operation
          schema:
            $ref: "#/definitions/Symbol"
  /api/vocabulary/{key}/symbol:
    get:
      tags:
      - Vocabulary
      description: Get all symbols in a vocabulary
      operationId: getSymbols
      parameters:
      - in: path
        name: key
        description: The key of an existing Vocabulary
        required: true
        type: string
      responses:
        200:
          description: Successful operation
          schema:
            $ref: "#/definitions/ArrayOfSymbol"
  /api/vocabulary/{key}/close:
    post:
      tags:
      - Vocabulary
      description: Declare this vocabulary closed, no more symbols will be added, description lengths will be determined accordingly
      operationId: closeVocabulary
      parameters:
      - in: path
        name: key
        description: The key of an existing Vocabulary
        required: true
        type: string
      responses:
        200:
          description: Successful operation
  /api/vocabulary/{key}/remain-open:
    post:
      tags:
      - Vocabulary
      description: Declare that this vocabulary remains open, more symbols can be added later, description lengths will be determined accordingly
      operationId: declareVocabularyOpen
      parameters:
      - in: path
        name: key
        description: The key of an existing Vocabulary
        required: true
        type: string
      responses:
        200:
          description: Successful operation
  /api/vocabulary-upload:
    post:
      tags:
      - Vocabulary
      description: Upload vocabulary
      operationId: uploadVocabulary
      consumes:
      - multipart/form-data
      parameters:
      - in: formData
        name: data
        description: The data to upload
        type: file
      responses:
        200:
          description: Successful operation
  /api/admin/unique-bucket/clean-existing-keys:
    post:
      tags:
      - UniqueKeys
      description: Move existing keys to their respective child buckets
      operationId: cleanExistingKeys
      responses:
        200:
          description: Successful operation
  /api/admin/unique-bucket:
    get:
      tags:
      - UniqueKeys
      description: Describe all unique buckets
      operationId: describeUniqueBuckets
      responses:
        200:
          description: Successful operation
          schema:
            $ref: "#/definitions/ArrayOfUniqueBucket"
definitions:
  ModelNodeData:
    type: object
    properties:
      path:
        description: Symbol IDs most recent first
        type: array
        items:
          $ref: "#/definitions/SymbolReference"
      key:
        description: Unique readable representation
        type: string
      extensible:
        description: Flag to indicate whether child nodes for this node may be created
        type: boolean
  Symbol:
    type: object
    properties:
      vocabulary:
        type: string
      name:
        type: string
      ordinal:
        type: integer
        format: int64
      descriptionLength:
        type: integer
  ArrayOfSymbol:
    type: array
    items:
      $ref: "#/definitions/Symbol"
  SymbolReference:
    type: object
    properties:
      vocabulary:
        type: string
      ordinal:
        type: integer
        format: int64
  UniqueBucket:
    type: object
    properties:
      id:
        type: string
      fullKeyPrefix:
        type: string
      keyPrefix:
        type: string
      parentId:
        type: string
  ArrayOfUniqueBucket:
    type: array
    items:
      $ref: "#/definitions/UniqueBucket"
  ArrayOfString:
    type: array
    items:
      type: string
securityDefinitions:
  STACK_auth:
    type: oauth2
    authorizationUrl: /oauth/authorize
    flow: implicit
    scopes:
      write:LeiaLearns: modify Leia Learns in your account
      read:LeiaLearns: read your Leia Learns
  api_key:
    type: apiKey
    name: api_key
    in: header